refactor: Validator 클래스 분리
